mapper, mapperImpl 제외 및 entity에 통합 travel #38
