+++ src/test/java/com/travel/travel_project/admin/travel/TravelRepositoryTest.java
@@ -1,7 +1,5 @@
 package com.travel.travel_project.admin.travel;
 
-import com.travel.travel_project.api.travel.mapper.group.TravelGroupMapper;
-import com.travel.travel_project.api.travel.mapper.review.TravelReviewMapper;
 import com.travel.travel_project.domain.travel.TravelDTO;
 import com.travel.travel_project.domain.travel.TravelEntity;
 import com.travel.travel_project.api.travel.TravelRepository;
@@ -36,6 +34,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
@@ -656,7 +655,7 @@ class TravelRepositoryTest {
                 .travelCode(1)
                 .travelTitle("여행지 테스트").travelDescription("여행지 테스트").favoriteCount(1).viewCount(0)
                 .travelAddress("인천광역시 서구").travelZipCode("123-456").visible("Y").popular(false)
-                .reviewList(TravelReviewMapper.INSTANCE.toEntityList(reviewList))
+                .reviewList(reviewList.stream().map(TravelReviewEntity::toEntity).collect(Collectors.toList()))
                 .build();
 
         // when
@@ -714,7 +713,7 @@ class TravelRepositoryTest {
                 .travelCode(1)
                 .travelTitle("여행지 테스트").travelDescription("여행지 테스트").favoriteCount(1).viewCount(0)
                 .travelAddress("인천광역시 서구").travelZipCode("123-456").visible("Y").popular(false)
-                .reviewList(TravelReviewMapper.INSTANCE.toEntityList(reviewList))
+                .reviewList(reviewList.stream().map(TravelReviewEntity::toEntity).collect(Collectors.toList()))
                 .build();
 
         // when
@@ -781,7 +780,7 @@ class TravelRepositoryTest {
                 .travelCode(1)
                 .travelTitle("여행지 테스트").travelDescription("여행지 테스트").favoriteCount(1).viewCount(0)
                 .travelAddress("인천광역시 서구").travelZipCode("123-456").visible("Y").popular(false)
-                .reviewList(TravelReviewMapper.INSTANCE.toEntityList(reviewList))
+                .reviewList(reviewList.stream().map(TravelReviewEntity::toEntity).collect(Collectors.toList()))
                 .build();
 
         // when
@@ -929,7 +928,7 @@ class TravelRepositoryTest {
 
         travelRepository.updateTravelGroup(newTravelGroupEntity);
 
-        TravelGroupDTO newTravelGroupDTO = TravelGroupMapper.INSTANCE.toDto(newTravelGroupEntity);
+        TravelGroupDTO newTravelGroupDTO = TravelGroupEntity.toDto(newTravelGroupEntity);
 
         // when
         when(mockTravelRepository.findOneTravelGroup(newTravelGroupEntity.getIdx())).thenReturn(newTravelGroupDTO);
@@ -957,7 +956,7 @@ class TravelRepositoryTest {
                 .travelIdx(1L).groupName("서울모임").groupDescription("서울모임").visible("Y").build();
         em.persist(travelGroupEntity);
 
-        TravelGroupDTO travelGroupDTO = TravelGroupMapper.INSTANCE.toDto(travelGroupEntity);
+        TravelGroupDTO travelGroupDTO = TravelGroupEntity.toDto(travelGroupEntity);
 
         // when
         when(mockTravelRepository.findOneTravelGroup(travelGroupDTO.getIdx())).thenReturn(travelGroupDTO);
@@ -983,7 +982,7 @@ class TravelRepositoryTest {
                 .travelIdx(1L).groupName("서울모임").groupDescription("서울모임").visible("Y").build();
         em.persist(travelGroupEntity);
 
-        TravelGroupDTO travelGroupDTO = TravelGroupMapper.INSTANCE.toDto(travelGroupEntity);
+        TravelGroupDTO travelGroupDTO = TravelGroupEntity.toDto(travelGroupEntity);
 
         TravelGroupUserEntity travelGroupUserEntity = TravelGroupUserEntity.builder()
                 .userIdx(1L).groupIdx(travelGroupDTO.getIdx()).build();
@@ -1003,7 +1002,7 @@ class TravelRepositoryTest {
                 .travelIdx(1L).groupName("서울모임").groupDescription("서울모임").visible("Y").build();
         em.persist(travelGroupEntity);
 
-        TravelGroupDTO travelGroupDTO = TravelGroupMapper.INSTANCE.toDto(travelGroupEntity);
+        TravelGroupDTO travelGroupDTO = TravelGroupEntity.toDto(travelGroupEntity);
 
         TravelGroupUserEntity travelGroupUserEntity = TravelGroupUserEntity.builder()
                 .userIdx(1L).groupIdx(travelGroupDTO.getIdx()).build();
