+++ src/test/java/com/travel/travel_project/admin/travel/TravelServiceTest.java
@@ -1,7 +1,5 @@
 package com.travel.travel_project.admin.travel;
 
-import com.travel.travel_project.api.travel.mapper.group.TravelGroupMapper;
-import com.travel.travel_project.api.travel.mapper.review.TravelReviewMapper;
 import com.travel.travel_project.domain.travel.TravelDTO;
 import com.travel.travel_project.domain.travel.TravelEntity;
 import com.travel.travel_project.api.travel.TravelService;
@@ -36,6 +34,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
@@ -505,7 +504,7 @@ class TravelServiceTest {
                 .travelCode(1)
                 .travelTitle("여행지 테스트").travelDescription("여행지 테스트").favoriteCount(1).viewCount(0)
                 .travelAddress("인천광역시 서구").travelZipCode("123-456").visible("Y").popular(false)
-                .reviewList(TravelReviewMapper.INSTANCE.toEntityList(reviewList))
+                .reviewList(reviewList.stream().map(TravelReviewEntity::toEntity).collect(Collectors.toList()))
                 .build();
 
         // when
@@ -563,7 +562,7 @@ class TravelServiceTest {
                 .travelCode(1)
                 .travelTitle("여행지 테스트").travelDescription("여행지 테스트").favoriteCount(1).viewCount(0)
                 .travelAddress("인천광역시 서구").travelZipCode("123-456").visible("Y").popular(false)
-                .reviewList(TravelReviewMapper.INSTANCE.toEntityList(reviewList))
+                .reviewList(reviewList.stream().map(TravelReviewEntity::toEntity).collect(Collectors.toList()))
                 .build();
 
         // when
@@ -630,7 +629,7 @@ class TravelServiceTest {
                 .travelCode(1)
                 .travelTitle("여행지 테스트").travelDescription("여행지 테스트").favoriteCount(1).viewCount(0)
                 .travelAddress("인천광역시 서구").travelZipCode("123-456").visible("Y").popular(false)
-                .reviewList(TravelReviewMapper.INSTANCE.toEntityList(reviewList))
+                .reviewList(reviewList.stream().map(TravelReviewEntity::toEntity).collect(Collectors.toList()))
                 .build();
 
         // when
@@ -778,7 +777,7 @@ class TravelServiceTest {
 
         travelService.updateTravelGroup(newTravelGroupEntity);
 
-        TravelGroupDTO newTravelGroupDTO = TravelGroupMapper.INSTANCE.toDto(newTravelGroupEntity);
+        TravelGroupDTO newTravelGroupDTO = TravelGroupEntity.toDto(newTravelGroupEntity);
 
         // when
         when(mockTravelService.findOneTravelGroup(newTravelGroupEntity.getIdx())).thenReturn(newTravelGroupDTO);
@@ -806,7 +805,7 @@ class TravelServiceTest {
                 .travelIdx(1L).groupName("서울모임").groupDescription("서울모임").visible("Y").build();
         em.persist(travelGroupEntity);
 
-        TravelGroupDTO travelGroupDTO = TravelGroupMapper.INSTANCE.toDto(travelGroupEntity);
+        TravelGroupDTO travelGroupDTO = TravelGroupEntity.toDto(travelGroupEntity);
 
         // when
         when(mockTravelService.findOneTravelGroup(travelGroupDTO.getIdx())).thenReturn(travelGroupDTO);
@@ -832,7 +831,7 @@ class TravelServiceTest {
                 .travelIdx(1L).groupName("서울모임").groupDescription("서울모임").visible("Y").build();
         em.persist(travelGroupEntity);
 
-        TravelGroupDTO travelGroupDTO = TravelGroupMapper.INSTANCE.toDto(travelGroupEntity);
+        TravelGroupDTO travelGroupDTO = TravelGroupEntity.toDto(travelGroupEntity);
 
         TravelGroupUserEntity travelGroupUserEntity = TravelGroupUserEntity.builder()
                 .userIdx(1L).groupIdx(travelGroupDTO.getIdx()).build();
@@ -852,7 +851,7 @@ class TravelServiceTest {
                 .travelIdx(1L).groupName("서울모임").groupDescription("서울모임").visible("Y").build();
         em.persist(travelGroupEntity);
 
-        TravelGroupDTO travelGroupDTO = TravelGroupMapper.INSTANCE.toDto(travelGroupEntity);
+        TravelGroupDTO travelGroupDTO = TravelGroupEntity.toDto(travelGroupEntity);
 
         TravelGroupUserEntity travelGroupUserEntity = TravelGroupUserEntity.builder()
                 .userIdx(1L).groupIdx(travelGroupDTO.getIdx()).build();
