+++ src/test/java/com/travel/travel_project/api/user/UserServiceTest.java
@@ -1,6 +1,5 @@
 package com.travel.travel_project.api.user;
 
-import com.travel.travel_project.api.travel.mapper.schedule.TravelScheduleMapper;
 import com.travel.travel_project.domain.travel.schedule.TravelScheduleDTO;
 import com.travel.travel_project.domain.travel.schedule.TravelScheduleEntity;
 import com.travel.travel_project.domain.user.UserDTO;
@@ -408,7 +407,7 @@ class UserServiceTest {
                 .build();
 
         em.persist(travelScheduleEntity);
-        TravelScheduleDTO travelScheduleDTO = TravelScheduleMapper.INSTANCE.toDto(travelScheduleEntity);
+        TravelScheduleDTO travelScheduleDTO = TravelScheduleEntity.toDto(travelScheduleEntity);
 
         // when
         when(mockUserService.findOneUserSchedule(travelScheduleDTO.getUserIdx(), travelScheduleDTO.getIdx())).thenReturn(travelScheduleDTO);
